requirements.txt
flask==3.1.0
google-generativeai==0.8.4
python-dotenv==1.1.0
gunicorn==23.0.0
Pillow==11.1.0
